auth_service/api/views.py
from django.shortcuts import render

# Create your views here.

from rest_framework import generics, permissions
from .serializers import RegisterSerializer
from rest_framework.decorators import api_view, permission_classes
from rest_framework.permissions import IsAuthenticated
from rest_framework.response import Response

class RegisterView(generics.CreateAPIView):
    serializer_class = RegisterSerializer
    permission_classes = [permissions.AllowAny]

@api_view(['GET'])
@permission_classes([IsAuthenticated])
def perfil(request):
    return Response({'username': request.user.username, 'email': request.user.email})
