requirements.txt
asgiref==3.9.1
certifi==2025.8.3
chardet==3.0.4
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==5.2.6
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.19.1
fonttools==4.60.0
fsspec==2025.9.0
googletrans==4.0.0rc1
gTTS==2.5.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==11.3.0
polars==1.33.1
psutil==7.1.0
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.5
rfc3986==1.5.0
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.202
ultralytics-thop==2.0.17
urllib3==2.5.0
